Close out the leftovers the merge left behind
**``enablement.stack_actions`` deleted.** Declared, read by the breakdown
collector, and published in the session-breakdown schema -- and written by
nothing outside two test fixtures, so ``have_actions`` was permanently
False and the field could never appear in a real session. Its three
siblings (``active_runtime``, ``attempt_runtimes``, ``kept_stack_action``)
each have a writer; this one was an unfinished idea advertised as a
contract. ``_stack_action_summary`` stays -- ``kept_stack_action`` uses it.

**The enablement package no longer reaches back into the source arm.**
``_coerce_needs_gpu`` / ``_framework_gpu_params`` /
``_framework_authoring_lanes_ttl`` decide GPU params and lane tiering for
any Coordinator-dispatched specialist; they lived on the source arm only
because that is where the first caller was. Moved to ``orchestrator.gpu_lanes``
as their own collaborator, so extracting enablement actually finished.

The move surfaced how easily that hop hides a fault: the relative import of
``is_multi_node`` was one level too deep in the new module, and the broad
``except`` around the probe swallowed the ImportError and reported
single-node -- which would have handed a multi-node run the whole machine.
``test_framework_gpu_params_empty_on_multi_node`` caught it. Depth fixed,
and the probe failure now logs at warning rather than debug.

**Retired-reason references.** The decision-priority prompt still named
``explore_no_more_leverage`` and ``framework_agent_plateau``; it now
describes the two-arm rule. The report glossary gains the ``optimize_*``
reasons and keeps the old ones so a report over an archived session still
explains what it is reading.
