+++ src/hyperloom/orchestrator/prompts/prompt_builder.py
@@ -611,12 +611,11 @@ def _section_decision_framework(*, kernel_enabled: bool, phase: str = "", transp
             "   ``pct`` share. In EXPLORE a ``force_exit`` line adds",
             "   ``phase_remaining_pct``; as it falls toward ``pct_thresh``,",
             "   prefer lower-cost / known-good actions (explore over kernel_opt).",
-            "   The Plateau advisory block is informational only for KERNEL;",
-            "   a detected FRAMEWORK_AGENT or EXPLORE plateau, by contrast,",
-            "   deterministically advances the phase. EXPLORE advances",
-            "   to KERNEL_AGENT (``reason=explore_no_more_leverage``) and",
-            "   FRAMEWORK_AGENT to EXPLORE (``reason=framework_agent_plateau``)",
-            "   at the next phase-compute. When you judge the current phase exhausted,",
+            "   The Plateau advisory block is informational only for KERNEL. In",
+            "   OPTIMIZE it reports each arm separately: BOTH arms dry advances",
+            "   to KERNEL_AGENT (``reason=optimize_no_more_leverage``) at the",
+            "   next phase-compute, while one arm dry means work the other.",
+            "   When you judge the current phase exhausted,",
             "   emit ``escalate_strategy_change{next_action_hint=",
             "   'skip_to_kernel' | 'skip_to_sweep' | 'skip_to_close'}`` (see",
             "   PHASE CONTRACT for the skip_to_close exception).",
