+++ src/hyperloom/orchestrator/state/_shared_state/enablement_round.py
@@ -67,8 +67,6 @@ class EnablementRound:
     revalidation_task_id: str = ""
     # Monotonically increasing counter for fresh revalidation idempotency keys.
     revalidation_generation: int = 0
-    # Attempt-scoped runtime acquisition state.
-    stack_actions: list = field(default_factory=list)
     active_runtime: dict = field(default_factory=dict)
     attempt_runtimes: list = field(default_factory=list)
     kept_stack_action: dict = field(default_factory=dict)
